+++ gdxsv_patch/src/main.c
@@ -49,8 +49,8 @@ struct sockaddr_t {
 };
 
 struct gdx_queue {
-    u16 head;
-    u16 tail;
+    u32 head;
+    u32 tail;
     u8 buf[GDX_QUEUE_SIZE];
 };
